refactor: add responsive file and refactor error page components
