+++ packages/author-profile/src/author-profile-error.web.js
@@ -1,42 +1,8 @@
 import React from "react";
-import { View } from "react-native";
 import PropTypes from "prop-types";
 import Image from "@times-components/image";
-import withResponsiveStyles from "@times-components/responsive-styles";
 import AuthorProfileListError from "./author-profile-list-error";
-
-const ErrorContainer = withResponsiveStyles(View, {
-  base: () => `
-    flex: 1;
-    justify-content: space-between;
-    flex-direction: column-reverse;
-    padding-top: 10%;
-    padding-left: 10%;
-    padding-right: 10%;
-  `,
-  mediumUp: () => `
-    padding-left: 15%;
-    padding-right: 15%;
-  `,
-  wideUp: () => `
-    flex-direction: row;
-  `
-});
-
-const ImageContainer = withResponsiveStyles(View, {
-  base: () => `
-    max-width: 75%;
-    width: 100%;
-    flex-basis: 50% !important;
-    align-self: center;
-  `,
-  mediumUp: () => `
-    max-width: 428px;
-  `,
-  wideUp: () => `
-    max-width: none;
-  `
-});
+import { ErrorContainer, ImageContainer } from "./styles/responsive";
 
 const AuthorProfileError = ({ refetch }) => (
   <ErrorContainer>
